refactor: MathPopup 폴더 생성해 코드 이동
