+++ src/features/math/components/MathPopups/MathCorrectPopup/index.tsx
