+++ src/features/math/components/MathPopups/MathSolutionPopup/index.tsx
