+++ src/features/math/components/MathPopups/index.ts
@@ -1,4 +1,1 @@
-export { default as MathCorrectPopup } from './MathCorrectPopup';
 export { default as MathPopupFactory } from './MathPopupFactory';
-export { default as MathSolutionPopup } from './MathSolutionPopup';
-export { default as MathWrongPopup } from './MathWrongPopup';
