+++ src/features/math/constants/popups.ts
@@ -1,6 +1,8 @@
 import type { ComponentProps, FunctionComponent } from 'react';
 
-import { MathCorrectPopup, MathSolutionPopup, MathWrongPopup } from '@/features/math/components/MathPopups';
+import MathCorrectPopup from '@/features/math/components/MathPopups/MathCorrectPopup';
+import MathSolutionPopup from '@/features/math/components/MathPopups/MathSolutionPopup';
+import MathWrongPopup from '@/features/math/components/MathPopups/MathWrongPopup';
 
 export const MATH_POPUPS = {
   correct: MathCorrectPopup as FunctionComponent<ComponentProps<typeof MathCorrectPopup>>,
